housekeeping/requirements.txt
asyncpg==0.29.0
cryptography==3.4.8
matplotlib==3.5.1
paramiko==3.2.0
Pillow==9.0.1
pwinput==1.0.3
python==3.9.12
PyYAML==6.0
scp==0.15
tk==0.1.0
pytz==2023.3
tzlocal==5.1
natsort==8.4.0